- Use of Last the version of Servoblaster Allows control of PWM signal in steps of 2 us instead of 10us. The effect on the stability is obvious. servo.cpp was modified accordingly
- Use of a pid control directly calculated on the Angle instead of both
rotationnal speed and Angle.

-Added an Init functionnality on the remote and in the Controller.
The initialization of the MPU6050 can now be done and checked prior
to starting the PID controller

+++ PILOT/pid.cpp
@@ -40,7 +40,7 @@ PID::PID()
   m_outmin = -20;
 }
 
-float PID::update_pid(float setpoint, float input, float dt)
+int PID::update_pid(float setpoint, float input, float dt)
 {
 
   //Computes error
@@ -72,7 +72,7 @@ float PID::update_pid(float setpoint, float input, float dt)
   //printf("setpt %f input %f outpt %f\n", m_Kp, m_Ki, m_Kd);
   //printf("setpt %f input %f outpt %f\n", setpoint, input, output);
 
-  return m_output;
+  return (int)m_output;
 }
 
 void PID::set_Kpid(float Kp,float Ki, float Kd)
